apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0
aioodbc==0.5.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
httpx==0.28.1
boto3==1.37.1
redis==5.2.1
celery==5.4.0
kombu==5.4.2
pyodbc==5.2.0
pymssql==2.3.2
sqlalchemy-pytds==0.3.0
pytz==2024.2
openai==1.58.1
pandas==2.2.3
openpyxl==3.1.5
lxml==5.3.0
orjson==3.10.12
structlog==24.4.0
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
azure-monitor-opentelemetry-exporter==1.0.0b33
sentry-sdk[fastapi]==2.19.2
websockets==14.1
zeep==4.3.2
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
